Refactored IAgreementsInfo part
