+++ src/frontend/src/components/public/PIAFormTabs/agreementsAndInformationBanks/helper/viewPIBTextInput.tsx
@@ -1,4 +1,4 @@
-import { ITextInputProps } from './AgreementsAndInformationBanks';
+import { ITextInputProps } from './IAgreementsInfo-interface';
 
 const PIBInputText = ({ pia }: ITextInputProps) => {
   return (
