+++ src/frontend/src/types/interfaces/pia-form.interface.ts
@@ -3,7 +3,7 @@ import { IAccuracyCorrectionAndRetention } from '../../components/public/PIAForm
 import { IAdditionalRisks } from '../../components/public/PIAFormTabs/additionalRisks/AdditionalRisks';
 import { ISecurityPersonalInformation } from '../../components/public/PIAFormTabs/securityPersonalInformation/security-personal-info-interface';
 import { ICollectionUseAndDisclosure } from '../../components/public/PIAFormTabs/collectionUseAndDisclosure/CollectionUseAndDisclosure';
-import { IAgreementsAndInformationBanks } from '../../components/public/PIAFormTabs/agreementsAndInformationBanks/AgreementsAndInformationBanks';
+import { IAgreementsAndInformationBanks } from '../../components/public/PIAFormTabs/agreementsAndInformationBanks/helper/IAgreementsInfo-interface';
 import { IStoringPersonalInformation } from '../../components/public/PIAFormTabs/storingPersonalInformation/interfaces';
 import { IPPQ } from '../../components/public/PIAFormTabs/ppq/interfaces';
 import { IReview } from '../../components/public/PIAFormTabs/review/interfaces';
